requirements.txt
absl-py==0.9.0
astor==0.8.1
astunparse==1.6.3
backcall==0.1.0
bleach==1.5.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
gast==0.2.2
google-auth==1.16.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
html5lib==0.9999999
idna==2.9
importlib-metadata==1.6.0
ipykernel==5.3.0
ipython==7.14.0
ipython-genutils==0.2.0
jedi==0.17.0
joblib==0.16.0
jupyter-client==6.1.3
jupyter-core==4.6.3
Keras==2.1.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.2.1
mock==4.0.2
numpy==1.18.4
oauthlib==3.1.0
opencv-python==4.2.0.34
opt-einsum==3.2.1
pandas==1.0.4
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
prompt-toolkit==3.0.5
protobuf==3.12.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.10.1
six==1.15.0
tensorboard==1.13.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow==1.13.2
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.2
tensorflow-tensorboard==0.1.8
termcolor==1.1.0
threadpoolctl==2.1.0
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.1.9
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
